refactor: class gltf utils
